Changed comparator and naming covnentions
Our frequencies should be doubles, so changed our comparator to accept doubles. Aslo changed some naming conventions in the main file, as well as added comments. Still on task 3, I think we need to explicitly implement left and right in our Node class.

+++ Comparato.java
@@ -2,6 +2,6 @@ import java.util.Comparator;
 
 public class Comparato implements Comparator<Node>{
     public int compare(Node x, Node y) {
-        return x.data - y.data;
+        return (int) (x.data - y.data);
     }
 }
